Move whitespace-only string test to test_tools.py
